webapisdefi

+++ WebApiParaiso/WebApiParaiso/Models/reservacion.cs
@@ -20,7 +20,7 @@ namespace WebApiParaiso.Models
 
         public int? idcliente { get; set; }
 
-        public int? total { get; set; }
+        public decimal? total { get; set; }
 
         public int? estado { get; set; }
